+++ vscode/extension/CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## 1.0.0
 
+* Enable using Go to Definition (F12) to open the file for an import.
 * Show the build error in the preview window.
 * Add error if a file tries to import itself.